+++ tests/can.rs
@@ -19,21 +19,13 @@ fn test_setup_can0() {
     let can_module = Module::new(Module0);
 
     // clear_cpu_endinit
-    report.expect_read(
-        SCU.wdtcpu()[0].wdtcpuycon0().ptr() as usize,
-        4,
-        0b11,
-    );
+    report.expect_read(SCU.wdtcpu()[0].wdtcpuycon0().ptr(), 4, 0b11);
 
     // clc read
-    report.expect_read(CAN0.clc().ptr() as usize, 4, 0b0);
+    report.expect_read(CAN0.clc().ptr(), 4, 0b0);
 
     // set_cpu_endinit
-    report.expect_read(
-        SCU.wdtcpu()[0].wdtcpuycon0().ptr() as usize,
-        4,
-        0b10,
-    );
+    report.expect_read(SCU.wdtcpu()[0].wdtcpuycon0().ptr(), 4, 0b10);
 
     let mut can_module = can_module.enable();
 
@@ -47,71 +39,71 @@ fn test_setup_can0() {
     };
 
     // mcr read for set clock source
-    report.expect_read(CAN0.mcr().ptr() as usize, 4, 0b0);
-    report.expect_read(CAN0.mcr().ptr() as usize, 4, 0b11);
+    report.expect_read(CAN0.mcr().ptr(), 4, 0b0);
+    report.expect_read(CAN0.mcr().ptr(), 4, 0b11);
 
     // cccr for enable configuration change
     // read
-    report.expect_read(CAN0.n()[0].cccri().ptr() as usize, 4, 0b1);
+    report.expect_read(CAN0.n()[0].cccri().ptr(), 4, 0b1);
 
     // modify
-    report.expect_read(CAN0.n()[0].cccri().ptr() as usize, 4, 0b1);
+    report.expect_read(CAN0.n()[0].cccri().ptr(), 4, 0b1);
 
     // read
-    report.expect_read(CAN0.n()[0].cccri().ptr() as usize, 4, 0b1);
+    report.expect_read(CAN0.n()[0].cccri().ptr(), 4, 0b1);
 
     // modify
-    report.expect_read(CAN0.n()[0].cccri().ptr() as usize, 4, 0b1);
+    report.expect_read(CAN0.n()[0].cccri().ptr(), 4, 0b1);
 
     // read
-    report.expect_read(CAN0.n()[0].cccri().ptr() as usize, 4, 0b0);
+    report.expect_read(CAN0.n()[0].cccri().ptr(), 4, 0b0);
 
     // modify
-    report.expect_read(CAN0.n()[0].cccri().ptr() as usize, 4, 0b0);
+    report.expect_read(CAN0.n()[0].cccri().ptr(), 4, 0b0);
 
     // read
-    report.expect_read(CAN0.n()[0].cccri().ptr() as usize, 4, 0b1);
+    report.expect_read(CAN0.n()[0].cccri().ptr(), 4, 0b1);
 
     // modify
-    report.expect_read(CAN0.n()[0].cccri().ptr() as usize, 4, 0b1);
+    report.expect_read(CAN0.n()[0].cccri().ptr(), 4, 0b1);
 
     // ccucon1 for get_mcan_frequency for configure_baud_rate
     report.expect_read(
-        SCU.ccucon1().ptr() as usize,
+        SCU.ccucon1().ptr(),
         4,
         0b0010_0001_0001_0001_0000_0010_0001_0010,
     );
 
     // ccucon0 read for get_source_frequency for get_mcan_frequency
     report.expect_read(
-        SCU.ccucon0().ptr() as usize,
+        SCU.ccucon0().ptr(),
         4,
         0b0001_0111_0010_0011_0000_0001_0001_0011,
     );
 
     // syspllcon0 read for get_osc_frequency for get_per_pll_frequency1 for get_source_frequency
     report.expect_read(
-        SCU.syspllcon0().ptr() as usize,
+        SCU.syspllcon0().ptr(),
         4,
         0b0100_0000_0000_0001_0011_1010_0000_0000,
     );
 
     // perpllcon0 read for get_per_pll_frequency1 for get_source_frequency
-    report.expect_read(SCU.perpllcon0().ptr() as usize, 4, 0b1_0011_1111_0000_0000);
+    report.expect_read(SCU.perpllcon0().ptr(), 4, 0b1_0011_1111_0000_0000);
 
     // perpllcon1 read for get_per_pll_frequency1 for get_source_frequency
-    report.expect_read(SCU.perpllcon1().ptr() as usize, 4, 0b1_0000_0001);
+    report.expect_read(SCU.perpllcon1().ptr(), 4, 0b1_0000_0001);
 
     // ccucon1 for get_source_frequency
     report.expect_read(
-        SCU.ccucon1().ptr() as usize,
+        SCU.ccucon1().ptr(),
         4,
         0b10_0001_0001_0001_0000_0010_0001_0010,
     );
 
     // nbtp0 for set_nominal_bit_timing
     report.expect_read(
-        CAN0.n()[0].nbtpi().ptr() as usize,
+        CAN0.n()[0].nbtpi().ptr(),
         4,
         0b110_0000_0000_0000_1010_0000_0011,
     );
@@ -121,25 +113,25 @@ fn test_setup_can0() {
         .expect("Cannot take can node");
 
     // txesc0 for set_tx_buffer_data_field_size for setup_tx
-    report.expect_read(CAN0.n()[0].tx().txesci().ptr() as usize, 4, 0b0);
+    report.expect_read(CAN0.n()[0].tx().txesci().ptr(), 4, 0b0);
 
     // txbc0 for set_tx_buffer_start_address for setup_tx
-    report.expect_read(CAN0.n()[0].tx().txbci().ptr() as usize, 4, 0b0);
+    report.expect_read(CAN0.n()[0].tx().txbci().ptr(), 4, 0b0);
 
     // txbc0 for set_dedicated_tx_buffers_number for setup_tx
-    report.expect_read(CAN0.n()[0].tx().txbci().ptr() as usize, 4, 0b100_0100_0000);
+    report.expect_read(CAN0.n()[0].tx().txbci().ptr(), 4, 0b100_0100_0000);
 
     // txbtie0 for enable_tx_buffer_transmission_interrupt for setup_tx
-    report.expect_read(CAN0.n()[0].tx().txbtiei().ptr() as usize, 4, 0b0);
-    report.expect_read(CAN0.n()[0].tx().txbtiei().ptr() as usize, 4, 0b1);
+    report.expect_read(CAN0.n()[0].tx().txbtiei().ptr(), 4, 0b0);
+    report.expect_read(CAN0.n()[0].tx().txbtiei().ptr(), 4, 0b1);
 
     // txefc0 for set_tx_event_fifo_start_address for setup_tx
-    report.expect_read(CAN0.n()[0].tx().txefci().ptr() as usize, 4, 0b0);
+    report.expect_read(CAN0.n()[0].tx().txefci().ptr(), 4, 0b0);
     // txefc0 for set_tx_event_fifo_size for setup_tx
-    report.expect_read(CAN0.n()[0].tx().txefci().ptr() as usize, 4, 0b100_0000_0000);
+    report.expect_read(CAN0.n()[0].tx().txefci().ptr(), 4, 0b100_0000_0000);
 
     // cccr0 for set_frame_mode for setup_tx
-    report.expect_read(CAN0.n()[0].cccri().ptr() as usize, 4, 0b11);
+    report.expect_read(CAN0.n()[0].cccri().ptr(), 4, 0b11);
 
     node.setup_tx(&TxConfig {
         mode: TxMode::DedicatedBuffers,
@@ -152,36 +144,36 @@ fn test_setup_can0() {
     });
 
     // rxesc0 for set_rx_buffer_data_field_size for setup_rx
-    report.expect_read(CAN0.n()[0].rx().rxesci().ptr() as usize, 4, 0b0);
+    report.expect_read(CAN0.n()[0].rx().rxesci().ptr(), 4, 0b0);
 
     // rxbc0 for set_rx_buffer_start_address for setup_rx
-    report.expect_read(CAN0.n()[0].rx().rxbci().ptr() as usize, 4, 0b0);
+    report.expect_read(CAN0.n()[0].rx().rxbci().ptr(), 4, 0b0);
 
     // rxesc0 for set_rx_fifo0_data_field_size for setup_rx
-    report.expect_read(CAN0.n()[0].rx().rxesci().ptr() as usize, 4, 0b0);
+    report.expect_read(CAN0.n()[0].rx().rxesci().ptr(), 4, 0b0);
 
     // rxf0c0 for set_rx_fifo0_start_address
-    report.expect_read(CAN0.n()[0].rx().rxf0ci().ptr() as usize, 4, 0b0);
+    report.expect_read(CAN0.n()[0].rx().rxf0ci().ptr(), 4, 0b0);
 
     // rxf0c0 for set_rx_fifo0_size
-    report.expect_read(CAN0.n()[0].rx().rxf0ci().ptr() as usize, 4, 0b1_0000_0000);
+    report.expect_read(CAN0.n()[0].rx().rxf0ci().ptr(), 4, 0b1_0000_0000);
 
     // rxf0c0 for set_rx_fifo0_operating_mode
     report.expect_read(
-        CAN0.n()[0].rx().rxf0ci().ptr() as usize,
+        CAN0.n()[0].rx().rxf0ci().ptr(),
         4,
         0b0100_0000_0001_0000_0000,
     );
 
     // rxf0c0 for set_rx_fifo0_watermark_level
     report.expect_read(
-        CAN0.n()[0].rx().rxf0ci().ptr() as usize,
+        CAN0.n()[0].rx().rxf0ci().ptr(),
         4,
         0b0100_0000_0001_0000_0000,
     );
 
     // cccr0 for set_frame_mode for setup_tx
-    report.expect_read(CAN0.n()[0].cccri().ptr() as usize, 4, 0b11);
+    report.expect_read(CAN0.n()[0].cccri().ptr(), 4, 0b11);
 
     node.setup_rx(RxConfig {
         mode: RxMode::SharedFifo0,
@@ -200,47 +192,31 @@ fn test_setup_can0() {
     });
 
     // clear_cpu_endinit
-    report.expect_read(
-        SCU.wdtcpu()[0].wdtcpuycon0().ptr() as usize,
-        4,
-        0b11,
-    );
+    report.expect_read(SCU.wdtcpu()[0].wdtcpuycon0().ptr(), 4, 0b11);
 
     // set_cpu_endinit
-    report.expect_read(
-        SCU.wdtcpu()[0].wdtcpuycon0().ptr() as usize,
-        4,
-        0b10,
-    );
+    report.expect_read(SCU.wdtcpu()[0].wdtcpuycon0().ptr(), 4, 0b10);
 
-    report.expect_read(CAN0.n()[0].npcri().ptr() as usize, 4, 0b0);
+    report.expect_read(CAN0.n()[0].npcri().ptr(), 4, 0b0);
 
     // clear_cpu_endinit
-    report.expect_read(
-        SCU.wdtcpu()[0].wdtcpuycon0().ptr() as usize,
-        4,
-        0b11,
-    );
+    report.expect_read(SCU.wdtcpu()[0].wdtcpuycon0().ptr(), 4, 0b11);
 
     // set_cpu_endinit
-    report.expect_read(
-        SCU.wdtcpu()[0].wdtcpuycon0().ptr() as usize,
-        4,
-        0b10,
-    );
+    report.expect_read(SCU.wdtcpu()[0].wdtcpuycon0().ptr(), 4, 0b10);
 
     node.setup_pins(Some(&Pins {
         tx: PIN_TX_0_0_P20_8,
         rx: PIN_RX_0_0_P20_7,
     }));
 
-    report.expect_read(CAN0.n()[0].grint2i().ptr() as usize, 4, 0b0);
+    report.expect_read(CAN0.n()[0].grint2i().ptr(), 4, 0b0);
 
-    report.expect_read(SRC.can().can()[0].canxinty()[1].ptr() as usize, 4, 0b0);
-    report.expect_read(SRC.can().can()[0].canxinty()[1].ptr() as usize, 4, 0b0);
-    report.expect_read(SRC.can().can()[0].canxinty()[1].ptr() as usize, 4, 0b0);
+    report.expect_read(SRC.can().can()[0].canxinty()[1].ptr(), 4, 0b0);
+    report.expect_read(SRC.can().can()[0].canxinty()[1].ptr(), 4, 0b0);
+    report.expect_read(SRC.can().can()[0].canxinty()[1].ptr(), 4, 0b0);
 
-    report.expect_read(CAN0.n()[0].iei().ptr() as usize, 4, 0b0);
+    report.expect_read(CAN0.n()[0].iei().ptr(), 4, 0b0);
 
     node.setup_interrupt(&NodeInterruptConfig {
         interrupt_group: InterruptGroup::Rxf0n,
@@ -250,10 +226,10 @@ fn test_setup_can0() {
         tos: Tos::Cpu0,
     });
 
-    report.expect_read(CAN0.n()[0].cccri().ptr() as usize, 4, 0b0);
-    report.expect_read(CAN0.n()[0].cccri().ptr() as usize, 4, 0b0);
-    report.expect_read(CAN0.n()[0].cccri().ptr() as usize, 4, 0b0);
-    report.expect_read(CAN0.n()[0].cccri().ptr() as usize, 4, 0b0);
+    report.expect_read(CAN0.n()[0].cccri().ptr(), 4, 0b0);
+    report.expect_read(CAN0.n()[0].cccri().ptr(), 4, 0b0);
+    report.expect_read(CAN0.n()[0].cccri().ptr(), 4, 0b0);
+    report.expect_read(CAN0.n()[0].cccri().ptr(), 4, 0b0);
 
     #[allow(unused_variables)]
     let node = node.lock_configuration();
